Fix minification script to account for embedded CSS in JS.
Apparently, pad.js contains CSS files embedded as strings.

Etherpad normally performs minification at runtime, and does some crazy loopback stuff to its own minification engine to cause these embeds themselves to be minified.

Sandstorm prefers minification in advance, via this script. But the script didn't account for embedded strings, so they didn't get minified. This not only meant that pad.js was larger than it was supposed to be, but it broke @import statements inside the CSS blobs.

This used not to matter because there were no @imports. But in 1.6.1, CSS for bullet/numbered lists was moved into a separate file and @imported. So, it broke.

Fixes https://github.com/sandstorm-io/sandstorm/issues/2881